+++ Calculator/Calculator/Controller/CalculatorViewController.swift
@@ -200,12 +200,3 @@ class CalculatorViewController: UIViewController {
         return numberFormatter.string(for: Double(input)) ?? "0"
     }
 }
-
-extension UIScrollView {
-    func scrollToBottom() {
-        self.layoutIfNeeded()
-        let bottomOffset = CGPoint(x: 0, y: contentSize.height - bounds.size.height)
-        setContentOffset(bottomOffset, animated: true)
-    }
-}
-
